+++ udata_front/theme/gouvfr/datagouv-components/src/components/Owner/OwnerTypeIcon.vue
@@ -1,11 +1,11 @@
 <template>
-  <Vicon :name="icon"/>
+  <Vicon v-if="icon" :name="icon"/>
 </template>
 <script setup lang="ts">
 import { computed } from "vue";
 import { OhVueIcon as Vicon, addIcons } from "oh-vue-icons";
 import { RiBankLine, RiBuilding2Line, RiCommunityLine, RiGovernmentLine, RiUserLine } from "oh-vue-icons/icons";
-import { ASSOCIATION, COMPANY, LOCAL_AUTHORITY, PUBLIC_SERVICE, USER, type OrganizationTypes, type UserType } from "../../composables/organizations/useOrganizationType";
+import { ASSOCIATION, COMPANY, LOCAL_AUTHORITY, OTHER, PUBLIC_SERVICE, USER, type OrganizationTypes, type UserType } from "../../composables/organizations/useOrganizationType";
 import { throwOnNever } from "../../helpers/throwOnNever";
 
 addIcons(RiBankLine, RiBuilding2Line, RiCommunityLine, RiGovernmentLine, RiUserLine);
@@ -26,6 +26,8 @@ const icon = computed(() => {
       return "ri-building-2-line";
     case USER:
       return "ri-user-line";
+    case OTHER:
+      return "";
     default:
       return throwOnNever(props.type, `Unknown type ${props.type}`);
   }