+++ udata_front/theme/gouvfr/datagouv-components/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Current (in progress)
 
-- Nothing yet
+- Add new resources_dowloads metric to Dataset type [#474](https://github.com/datagouv/udata-front/pull/474)
+- Improve organization type logic [#475](https://github.com/datagouv/udata-front/pull/475)
 
 ## 1.1.2 (2024-08-01)