+++ udata_front/theme/gouvfr/datagouv-components/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Current (in progress)
 
+- Nothing yet
+
+## 1.2.0 (2024-08-21)
+
 - Add new resources_dowloads metric to Dataset type [#474](https://github.com/datagouv/udata-front/pull/474)
 - Improve organization type logic [#475](https://github.com/datagouv/udata-front/pull/475)
 - Update reuse card [#449](https://github.com/datagouv/udata-front/pull/449)